Move table code to result_table_manager - and testing to test_result_table_manager
