+++ tests/test_plot_manager.py
@@ -124,7 +124,6 @@ class TestPlotManager(trc.TestResultCollector):
 
         self._single_model_config = config
 
-        self._single_model_result_manager.create_tables()
         self._single_model_result_manager.compile_and_sort_results()
 
     def _create_multi_model_result_manager(self):
@@ -145,7 +144,6 @@ class TestPlotManager(trc.TestResultCollector):
 
         self._multi_model_config = config
 
-        self._multi_model_result_manager.create_tables()
         self._multi_model_result_manager.compile_and_sort_results()
 
     def _plot_manager_to_dict(self, plot_manager):
